Use Guice to create AzkabanExecutorServer for testing
- Single constructor
- Inject jetty Server instance from module
